refactor(planning): decompose PlanningManager into Facade + 2 Mixins
Extract timeline generation and feasibility validation into dedicated
mixins (TimelineMixin, ValidationMixin) following the established
Facade + Mixins pattern. CRUD methods and models stay in the facade.
