+++ magma_cycling/planning/manager/__init__.py
@@ -1,0 +1,1 @@
+"""PlanningManager mixin modules for god class decomposition."""
